Ported 333369152154ccd04b7906a5b47bc433ce962632 from corefx: SqlConnection.RetrieveStatistics should always return DictionaryEntry
To break the dependency on the non-generic collections,
SqlConnection.RetrieveStatistics was changed to return a
Hashtable-compatible subclass of Dictionary<TKey, TValue> instead of
returning Hashtable. However, a subtle behavior difference was missed.

Dictionary<TKey, TValue>'s implementation of IDictionary.GetEnumerator
returns a DictionaryEntry enumerator, which matches Hashtable. Since the
return type of RetrieveStatistics is IDictionary, the default behavior
of enumerating the result is the same as Hashtable.

However, if the result is cast from IDictionary to IEnumerable,
Dictionary<TKey, TValue>'s IEnumerable.GetEnumerator implementation will
return a KeyValuePair<TKey, TValue> enumerator, which does not match
Hashtable's behavior of returning a DictionaryEntry enumerator.

This change makes IEnumerable.GetEnumerator return a DictionaryEntry
enumerator to match Hashtable.